Reenable linter and code style standardization

+++ src/silo/query_engine/operators/union.cpp
@@ -1,7 +1,13 @@
 #include "silo/query_engine/operators/union.h"
 
-#include <roaring/roaring.hh>
+#include <algorithm>
+#include <cstddef>
+#include <iterator>
+#include <string>
+#include <utility>
 #include <vector>
+
+#include <roaring/roaring.hh>
 
 #include "silo/query_engine/operators/complement.h"
 #include "silo/query_engine/operators/operator.h"
@@ -16,7 +22,7 @@ Union::~Union() noexcept = default;
 
 std::string Union::toString() const {
    std::string res = "(" + children[0]->toString();
-   for (unsigned i = 1; i < children.size(); ++i) {
+   for (size_t i = 1; i < children.size(); ++i) {
       const auto& child = children[i];
       res += " | " + child->toString();
    }
@@ -29,10 +35,10 @@ Type Union::type() const {
 }
 
 OperatorResult Union::evaluate() const {
-   const unsigned size_of_children = children.size();
+   const uint32_t size_of_children = children.size();
    std::vector<const roaring::Roaring*> union_tmp(size_of_children);
    std::vector<OperatorResult> child_res(size_of_children);
-   for (unsigned i = 0; i < size_of_children; i++) {
+   for (uint32_t i = 0; i < size_of_children; i++) {
       child_res[i] = children[i]->evaluate();
       const roaring::Roaring& const_bitmap = *child_res[i];
       union_tmp[i] = &const_bitmap;
